Arreglando borrar tipo habitacion
Al borrar el tipo de habitación controlamos que ese tipo de habitaciones no tengan reservas activas y borramos los servicios adjuntos a estas

+++ Reservas/BD_habitaciones.php
@@ -151,50 +151,50 @@ function borrar_habitaciones($habitaciones_borrar) {
  * 
  * @param array $tipos_habitaciones_a_borrar Array de habitaciones a borrar
  */
-function borrar_tipo_habitaciones($tipos_habitaciones_a_borrar, $id_tipo) {
-
+function borrar_tipo_habitaciones($tipo_habitacion_a_borrar, $id_tipo) {
+    echo "Intentando borrar tipo $tipo_habitacion_a_borrar con id $id_tipo <br>";
     try {
         $base = conectar('admin');
         $base->beginTransaction();
 
-        for ($index = 0; $index < count($tipos_habitaciones_a_borrar); $index++) {
 
-//Borrando servicios dependientes del tipo de habitación
-//PENDIENTE
+
 //Borrando reservas dependientes del tipo de habitación
 //PENDIENTE
+//Borrando servicios adjuntos al tipo de habitación
+            $sentencia4 = $base->prepare("DELETE FROM habitacion_servicio WHERE tipo_habitacion=:tipo");
+            $sentencia4->bindParam(':tipo', $tipo_habitacion_a_borrar);
+            $sentencia4->execute();
 //Borrando las imagenes dependientes del tipo de habitación
 
-            $sentencia = $base->prepare("DELETE FROM imagenes_habitaciones WHERE id_tipo_habitacion=:id_tipo;");
+        $sentencia = $base->prepare("DELETE FROM imagenes_habitaciones WHERE id_tipo_habitacion=:id_tipo;");
 
-            $sentencia->bindParam(':id_tipo', $id_tipo[$index]);
+        $sentencia->bindParam(':id_tipo', $id_tipo);
 
-            $sentencia->execute();
+        $sentencia->execute();
 
 //Borrar imagenes del servidor
-            for ($i = 0; $i < 5; $i++) {
-                borrar_img_servidor((($id_tipo[$index] * 5) + 1) + $i);
-            }
+
+        for ($i = 0; $i < 5; $i++) {
+            borrar_img_servidor((($id_tipo * 5) + 1) + $i);
+        }
+
+
 //Borrando habitaciones del tipo indicado
-            $sentencia2 = $base->prepare("DELETE FROM habitaciones WHERE tipo_habitacion=:tipo_habitacion");
-            $sentencia2->bindParam(':tipo_habitacion', $tipos_habitaciones_a_borrar[$index]);
-            $sentencia2->execute();
+        $sentencia2 = $base->prepare("DELETE FROM habitaciones WHERE tipo_habitacion=:tipo_habitacion");
+        $sentencia2->bindParam(':tipo_habitacion', $tipo_habitacion_a_borrar);
+        $sentencia2->execute();
 
 //Borrando el tipo
-            $sentencia3 = $base->prepare("DELETE FROM tipo_habitaciones WHERE tipo_de_habitacion=:tipo_habitacion");
-            $sentencia3->bindParam(':tipo_habitacion', $tipos_habitaciones_a_borrar[$index]);
-            $sentencia3->execute();
+        $sentencia3 = $base->prepare("DELETE FROM tipo_habitaciones WHERE tipo_de_habitacion=:tipo_habitacion");
+        $sentencia3->bindParam(':tipo_habitacion', $tipo_habitacion_a_borrar);
+        $sentencia3->execute();
 
-            $msg = "listo";
-            echo "<script type='text/javascript'>alert($msg);</script>";
-            print 'Listo para ejecutar';
-            print_r("listo");
-            echo 'Listo';
+        $sentencia = null;
+        $sentencia2 = null;
+        $sentencia3 = null;
+//            $sentencia4 = null;
 
-            $sentencia = null;
-            $sentencia2 = null;
-            $sentencia3 = null;
-        }
         $base->commit();
         echo 'Datos borrados correctamente';
         $base = null;
@@ -210,7 +210,9 @@ function borrar_tipo_habitaciones($tipos_habitaciones_a_borrar, $id_tipo) {
  * @param type $index
  */
 function borrar_img_servidor($index) {
-    unlink('Imagenes_habitaciones/' . $index . '.jpg');
+    if (file_exists("'Imagenes_habitaciones/' . $index . '.jpg'")) {
+        unlink('Imagenes_habitaciones/' . $index . '.jpg');
+    }
 }
 
 /**
@@ -505,6 +507,19 @@ function consultar_datos_administradores($id_usuario) {
         $sentencia->bindParam(":id_usuario", $id_usuario);
         $sentencia->execute();
         $resultados = $sentencia->fetchAll();
+        return $resultados;
+    } catch (PDOException $e) {
+        print $e->getMessage();
+    }
+}
+
+function consultar_reservas_por_tipo($tipo) {
+    try {
+        $base = conectar('admin');
+        $sql = $base->prepare("SELECT * FROM habitaciones_reservas INNER JOIN habitaciones ON habitaciones_reservas.id_habitacion=habitaciones.id WHERE habitaciones.tipo_habitacion=:tipo");
+        $sql->bindParam(":tipo", $tipo);
+        $sql->execute();
+        $resultados = $sql->fetchAll();
         return $resultados;
     } catch (PDOException $e) {
         print $e->getMessage();
